Cleaning up files to use the new google cloud app
Some of the files were exposing vulnerable information so I had to set up a new Google Cloud application.
Now that I've done that, I created new keys and am keeping them locally so as not to expose them on Github.
This should be secure for now, but eventually I'm going to need to find a way to store them securely.
Doing some research...